% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyUROMOL.R
\name{IdentifyUROMOL}
\alias{IdentifyUROMOL}
\title{UROMOL2021 class of non-muscle-invasive bladder cancer (NMIBC) samples identified by methylation values}
\usage{
IdentifyUROMOL(x, minCor = 0.2)
}
\arguments{
\item{x}{data.frame with CpGs from HM450K or Infinium MethylationEPIC (Epic) in rows and samples to be classified in columns (accepting a dataframe with a single column namely one sample).}

\item{minCor}{numeric value specifying a minimal threshold for best Pearson's correlation between a sample's methylation profile and centroids profiles.
A sample showing no correlation above this threshold will remain unclassifed and prediction results will be set to NA. Default minCor value is 0.2.}
}
\value{
dataframe with classification results. The UROMOL_2021_class column returns the predicted class labels of the samples.
The cor_pval column returns the p-value associated with the Pearson's correlation between the sample and the nearest centroid.
The separationLevel gives a measure (ranging from 0 to 1) of how a sample is representative of its consensus class,
with 0 meaning the sample is too close to the other classes to be confidently assigned to one class label, and 1 meaning the sample is very representative of its class.
This separationLevel is measured as follows: (correlation to nearest centroid - correlation to second nearest centroid) / median difference of sample-to-centroid correlation.
The remaining four columns return the Pearson's correlation values for each sample and each centroid.
UROMOL_2021_class predictions are set to NA if the minCor condition is not verified.
}
\description{
This package performs a Pearson nearest-centroid classification according to the UROMOL2021 classes of NMIBC based on methylation values from Infinium HumanMethylation450 BeadChip (HM450K)
}
\note{
This classifier was built referenced to the published tool for the consensus classes of NMIBC:
Lindskrog, S.V., Prip, F., Lamy, P. et al. An integrated multi-omics analysis identifies prognostic molecular subtypes of non-muscle-invasive bladder cancer. Nat Commun 12, 2301 (2021). https://doi.org/10.1038/s41467-021-22465-w
}
\examples{
data(test_data)
IdentifyUROMOL(test_data, minCor = .2)

}
